requirements.txt
django==2.2.3
djangorestframework==3.10.2
psycopg2-binary==2.8.3
